+++ plantcv/geospatial/__init__.py
@@ -3,6 +3,10 @@ from plantcv.geospatial.read_geotif import read_geotif
 from plantcv.geospatial.transform_points import transform_points
 from plantcv.geospatial.transform_polygons import transform_polygons
 from plantcv.geospatial.points_to_geojson import points_to_geojson
+from plantcv.geospatial.points2roi import points2roi_circle
+from plantcv.geospatial.shapes_to_geojson import shapes_to_geojson
+from plantcv.geospatial import analyze
+from plantcv.geospatial import shapes
 from plantcv.geospatial.napari_grid import napari_grid
 from plantcv.geospatial.napari_polygon_grid import _lineintersect, napari_polygon_grid
 from plantcv.geospatial.center_grid_rois import center_grid_rois
@@ -15,8 +19,13 @@ __all__ = [
     "transform_points",
     "transform_polygons",
     "points_to_geojson",
+    "points2roi_circle",
+    "analyze",
+    "points2roi_circle",
+    "shapes_to_geojson",
+    "shapes",
     "napari_grid",
     "_lineintersect",
     "napari_polygon_grid",
     "center_grid_rois"
-    ]
+]