+++ plantcv/geospatial/read_geotif.py
@@ -160,11 +160,11 @@ def read_geotif(filename, bands="R,G,B", cropto=None):
                             img_data[:, :, [id_green]],
                             img_data[:, :, [id_red]]))
     # Gamma correction
-    if pseudo_rgb.dtype != 'uint8':
-        pseudo_rgb = pseudo_rgb.astype('float32') ** (1 / 2.2)
-        pseudo_rgb = pseudo_rgb * 255
-        pseudo_rgb = pseudo_rgb.astype('uint8')
-
+    # if pseudo_rgb.dtype != 'uint8':
+    #     pseudo_rgb = pseudo_rgb.astype('float32') ** (1 / 2.2)
+    #     pseudo_rgb = pseudo_rgb * 255
+    #     pseudo_rgb = pseudo_rgb.astype('uint8')
+    pseudo_rgb = pseudo_rgb.astype('uint8')
     # Make a Spectral_data instance before calculating a pseudo-rgb
     spectral_array = Spectral_data(array_data=img_data,
                                    max_wavelength=max(wavelengths, key=wavelengths.get),