+++ tests/conftest.py
@@ -14,20 +14,32 @@ class TestData:
         self.snapshot_dir = os.path.join(self.datadir, "snapshot_dir")
         # multispectral image
         self.cropped_tif = os.path.join(self.datadir, "615.tif")
+        # multispectral pickled
+        self.multi_pickled = os.path.join(self.datadir, "multi_img.spectral")
         # empty image
         self.empty_tif = os.path.join(self.datadir, "cropped_empty.tif")
         # rgb image
         self.rgb_tif = os.path.join(self.datadir, "rgb.tif")
+        # rgb pickled
+        self.rgb_pickled = os.path.join(self.datadir, "rgb_img.spectral")
         # multiPoints shapefilex
         self.pts_geojson = os.path.join(self.datadir, "test_pts.geojson")
         # points shapefilex
         self.single_pts_geojson = os.path.join(self.datadir, "single_test_pts.geojson")
         # polygon shapefile
         self.square_crop = os.path.join(self.datadir, "square_crop.geojson")
+         # polygon shapefile with "ID" in properties
+        self.geojson_with_id = os.path.join(self.datadir, "square_crop_with_id.geojson")
         # points shapefile
         self.point_crop = os.path.join(self.datadir, "point_crop.geojson")
         # multi polygon shapefile
         self.multipolygon = os.path.join(self.datadir, "multipolygon_fortests.geojson")
+        # epsg4326 points shapefile
+        self.epsg4326_geojson = os.path.join(self.datadir, "epsg4326points.geojson")
+        # plot bounds shapefile
+        self.plot_bounds = os.path.join(self.datadir, "docs_four_points.geojson")
+        # plot points shapefile
+        self.plot_points = os.path.join(self.datadir, "plot_points.geojson")
 
 @pytest.fixture(scope="session")
 def test_data():