+++ docs/changelog.md
@@ -2,9 +2,18 @@
 
 All notable changes to this project will be documented below.
 
+
+#### geospatial.analyze.coverage
+
+* v0.1dev: **geospatial.analyze.coverage**(*img, bin_mask, geojson*)
+
+#### geospatial.points2roi_circle
+
+* v0.1dev: rois = **geospatial.points2roi_circle**(*img, geojson, radius*)
+
 #### geospatial.points_to_geojson
 
-* v0.1dev: **geosptial.points_to_geojson**(*img, viewer, out_path*)
+* v0.1dev: **geospatial.points_to_geojson**(*img, viewer, out_path*)
 
 #### geospatial.center_grid_rois
 
@@ -22,6 +31,15 @@ All notable changes to this project will be documented below.
 
 * v0.1dev: spectral = **geospatial.read_geotif**(*filename, bands="B,G,R", cropto=None*)
 
+#### geospatial.shapes.grid
+
+* v0.1dev: cells = **geospatial.shapes.grid**(*img, field_corners_path, out_path, num_ranges, num_columns,
+         range_length, row_length, num_rows=1, range_spacing=0, column_spacing=0*)
+
+#### geospatial.shapes.flexible
+
+* v0.1dev: cells = **geospatial.shapes.flexible**(*img, field_corners_path, plot_geojson_path, out_path, range_length, row_length, num_rows=1*)
+
 #### geospatial.transform_points
 
 * v0.1dev: coord = **geospatial.transform_points**(*img, geojson*)
@@ -29,3 +47,7 @@ All notable changes to this project will be documented below.
 #### geospatial.transform_polygons
 
 * v0.1dev: coord = **geospatial.transform_polygons**(*img, geojson*)
+
+#### geospatial.shapes_to_geojson
+
+* v0.1dev: coord = **geospatial.shapes_to_geojson**(*img, viewer, out_path*)